update, fixing SQL issues

+++ dfe3_project/src/main/resources/library-schema.sql
@@ -2,7 +2,7 @@ drop table if exists 'library' CASCADE;
 create table
 	'library'
 	(
-		id integer AUTO_INCREMENT,
+		id bigint AUTO_INCREMENT,
 		title varchar(128),
 		author varchar(128),
 		genre varchar(128)
